requirements.txt
websocket-client
numpy
pyaudio